change transitionDuration from Int to Double

+++ app/src/main/java/com/immichframe/immichframe/MainActivity.kt
@@ -91,7 +91,7 @@ class MainActivity : AppCompatActivity() {
         val imageStretch: String,
         val margin: String,
         val interval: Int,
-        val transitionDuration: Int,
+        val transitionDuration: Double,
         val downloadImages: Boolean,
         val renewImagesDuration: Int,
         val showClock: Boolean,
